relocate utils files
