+++ cli/utils/resurrection.ts
@@ -1,6 +1,6 @@
 import { createLogger } from '../../src/utils/logger.js';
-import { getLastMemoryCid } from '../../src/agents/tools/utils/blockchain/agentMemoryContract.js';
-import { download } from '../../src/agents/tools/utils/dsn/dsnDownload.js';
+import { getLastMemoryCid } from '../../src/utils/blockchain/agentMemoryContract.js';
+import { download } from '../../src/utils/dsn/dsnDownload.js';
 import { existsSync, readFileSync, writeFileSync } from 'fs';
 import { join } from 'path';
 import { config } from '../../src/config/index.js';
