+++ src/utils/dsn/autoDriveApi.ts
@@ -1,5 +1,5 @@
 import { createAutoDriveApi } from '@autonomys/auto-drive';
-import { config } from '../../../../config/index.js';
+import { config } from '../../config/index.js';
 
 export const autoDriveApi = createAutoDriveApi({
   apiKey: config.autoDriveConfig.AUTO_DRIVE_API_KEY || '',
